input getRandNumbers
